DEVOPS-725: Print less logs on startup.
If a container already exists, `docker-compose logs -f` prints all
logs since the container booted up for the first time.
Especially if you are resuming a project, this can take several seconds to
output all of the container logs.

Would prefer to use --since but that only exists on `docker logs`